Make tt_tenfun a tensor method to more closely align to elemfun and MATLAB
