+++ docs/source/matlab/common.rst
@@ -56,7 +56,7 @@ Methods
 +-----------------+----------------------+------------------------------------------------------------------------+
 | ``subsref``     | ``__getitem__``      | ``X[index]``                                                           |
 +-----------------+----------------------+------------------------------------------------------------------------+
-| ``tenfun``      | ``tt_tenfun``        | e.g., ``pyttb.tt_tenfun(lambda x: x + 1, A)``                          |
+| ``tenfun``      | ``tenfun``           | ``X.tenfun(lambda x: x + 1)``                                          |
 +-----------------+----------------------+------------------------------------------------------------------------+
 | ``times``       | ``__mul__``          | ``X * Y``                                                              |
 +-----------------+----------------------+------------------------------------------------------------------------+
