+++ pyttb/tensor.py
@@ -8,9 +8,10 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Iterable
+from inspect import signature
 from itertools import combinations_with_replacement, permutations
 from math import factorial
-from typing import Any, Callable, List, Literal, Optional, Tuple, Union, overload
+from typing import Any, Callable, List, Literal, Optional, Tuple, Union, cast, overload
 
 import numpy as np
 import scipy.sparse.linalg
@@ -28,7 +29,6 @@ from pyttb.pyttb_utils import (
     tt_ind2sub,
     tt_sub2ind,
     tt_subsubsref,
-    tt_tenfun,
 )
 
 
@@ -782,7 +782,7 @@ class tensor:
         def logical_and(x, y):
             return np.logical_and(x, y).astype(dtype=x.dtype)
 
-        return tt_tenfun(logical_and, self, other)
+        return self.tenfun(logical_and, other)
 
     def logical_not(self) -> tensor:
         """
@@ -816,7 +816,7 @@ class tensor:
         def tensor_or(x, y):
             return np.logical_or(x, y).astype(x.dtype)
 
-        return tt_tenfun(tensor_or, self, other)
+        return self.tenfun(tensor_or, other)
 
     def logical_xor(self, other: Union[float, tensor]) -> tensor:
         """
@@ -837,7 +837,7 @@ class tensor:
         def tensor_xor(x, y):
             return np.logical_xor(x, y).astype(dtype=x.dtype)
 
-        return tt_tenfun(tensor_xor, self, other)
+        return self.tenfun(tensor_xor, other)
 
     def mask(self, W: tensor) -> np.ndarray:
         """
@@ -1756,6 +1756,169 @@ class tensor:
             return y
         assert False, "Invalid value for version; should be None, 1, or 2"
 
+    def tenfun(
+        self,
+        function_handle: Union[
+            Callable[[np.ndarray, np.ndarray], np.ndarray],
+            Callable[[np.ndarray], np.ndarray],
+        ],
+        *inputs: Union[
+            float,
+            int,
+            np.ndarray,
+            ttb.tensor,
+            ttb.ktensor,
+            ttb.ttensor,
+            ttb.sptensor,
+            ttb.sumtensor,
+        ],
+    ) -> ttb.tensor:
+        """
+        Apply a function to each element in a tensor or tensors
+
+        See :meth:`pyttb.tensor.tenfun_binary` and
+        :meth:`pyttb.tensor.tenfun_binary_unary` for supported
+        options.
+        """
+        assert callable(function_handle), "function_handle must be callable"
+
+        # Number of inputs for function handle
+        nfunin = len(signature(function_handle).parameters)
+
+        # Case I: Binary function
+        if len(inputs) == 1 and nfunin == 2:
+            # We manually inspected the function handle for the parameters
+            # maybe there is a more clever way to convince mypy
+            binary_function_handle = cast(
+                Callable[[np.ndarray, np.ndarray], np.ndarray], function_handle
+            )
+            Y = inputs[0]
+            if not isinstance(Y, (int, float)):
+                Y = self._tt_to_tensor(Y)
+            return self.tenfun_binary(binary_function_handle, Y)
+
+        # Convert inputs to tensors if they aren't already
+        # Allow inputs to be mutable in case of type conversion
+        input_tensors: list[Union[ttb.tensor]] = []
+        for an_input in inputs:
+            if not isinstance(
+                an_input,
+                (
+                    np.ndarray,
+                    ttb.tensor,
+                    ttb.ktensor,
+                    ttb.ttensor,
+                    ttb.sptensor,
+                    ttb.sumtensor,
+                ),
+            ):
+                assert (
+                    False
+                ), f"Invalid input to ten fun: {an_input} of type {type(an_input)}"
+            input_tensors.append(self._tt_to_tensor(an_input))
+
+        # Case II: Expects input to be matrix and applies operation on each columns
+        if nfunin != 1:
+            raise ValueError(
+                "Tenfun only supports binary and unary function handles but provided "
+                "function handle takes {nfunin} arguments."
+            )
+        unary_function_handle = cast(
+            Callable[[np.ndarray], np.ndarray], function_handle
+        )
+        return self.tenfun_unary(unary_function_handle, *input_tensors)
+
+    def tenfun_binary(
+        self,
+        function_handle: Callable[[np.ndarray, np.ndarray], np.ndarray],
+        other: Union[ttb.tensor, int, float],
+        first: bool = True,
+    ) -> ttb.tensor:
+        """Apply a binary operation to two tensors or a tensor and a scalar.
+
+        Parameters
+        ----------
+        function_handle: Function to apply.
+        other: Other input to the binary function.
+        first: Whether the tensor comes first in the method call (if ordering matters).
+
+        Example
+        -------
+        >>> add = lambda x, y: x + y
+        >>> t0 = ttb.tenones((2, 2))
+        >>> t1 = t0.tenfun_binary(add, t0)
+        >>> t1.isequal(t0 * 2)
+        True
+        >>> t2 = t0.tenfun_binary(add, 1)
+        >>> t2.isequal(t1)
+        True
+        """
+        X = self.data
+        if not isinstance(other, (float, int)):
+            Y = other.data
+        else:
+            Y = np.array(other)
+
+        if not first:
+            Y, X = X, Y
+        data = function_handle(X, Y)
+        Z = ttb.tensor(data, copy=False)
+        return Z
+
+    def tenfun_unary(
+        self, function_handle: Callable[[np.ndarray], np.ndarray], *inputs: ttb.tensor
+    ) -> ttb.tensor:
+        """Apply a unary operation to multiple tensors columnwise.
+
+        Example
+        -------
+        >>> tensor_max = lambda x: np.max(x, axis=0)
+        >>> data = np.array([[1, 2, 3], [4, 5, 6]])
+        >>> t0 = ttb.tensor(data)
+        >>> t1 = ttb.tensor(data)
+        >>> t2 = t0.tenfun_unary(tensor_max, t1)
+        >>> t2.isequal(t1)
+        True
+        """
+        sz = self.shape
+        for i, an_input in enumerate(inputs):
+            if isinstance(an_input, (float, int)):
+                assert False, f"Argument {i} is a scalar but expected a tensor"
+            elif sz != an_input.shape:
+                assert (
+                    False
+                ), f"Tensor {i} is not the same size as the first tensor input"
+        if len(inputs) == 0:
+            X = self.data
+            X = np.reshape(X, (1, -1))
+        else:
+            X = np.zeros((len(inputs) + 1, np.prod(sz)))
+            X[0, :] = np.reshape(self.data, (np.prod(sz)))
+            for i, an_input in enumerate(inputs):
+                X[i + 1, :] = np.reshape(an_input.data, (np.prod(sz)))
+        data = function_handle(X)
+        data = np.reshape(data, sz)
+        Z = ttb.tensor(data, copy=False)
+        return Z
+
+    def _tt_to_tensor(
+        self,
+        some_tensor: Union[
+            np.ndarray,
+            ttb.tensor,
+            ttb.ktensor,
+            ttb.ttensor,
+            ttb.sptensor,
+            ttb.sumtensor,
+        ],
+    ) -> ttb.tensor:
+        """Convert a variety of data structures to a dense tensor."""
+        if isinstance(some_tensor, np.ndarray):
+            return ttb.tensor(some_tensor)
+        elif isinstance(some_tensor, ttb.tensor):
+            return some_tensor
+        return some_tensor.to_tensor()
+
     def __setitem__(self, key, value):
         """
         Subscripted assignment for a tensor.
@@ -2048,7 +2211,7 @@ class tensor:
         def tensor_equality(x, y):
             return x == y
 
-        return tt_tenfun(tensor_equality, self, other)
+        return self.tenfun(tensor_equality, other)
 
     def __ne__(self, other):
         """
@@ -2080,7 +2243,7 @@ class tensor:
         def tensor_not_equal(x, y):
             return x != y
 
-        return tt_tenfun(tensor_not_equal, self, other)
+        return self.tenfun(tensor_not_equal, other)
 
     def __ge__(self, other):
         """
@@ -2112,7 +2275,7 @@ class tensor:
         def greater_or_equal(x, y):
             return x >= y
 
-        return tt_tenfun(greater_or_equal, self, other)
+        return self.tenfun(greater_or_equal, other)
 
     def __le__(self, other):
         """
@@ -2144,7 +2307,7 @@ class tensor:
         def less_or_equal(x, y):
             return x <= y
 
-        return tt_tenfun(less_or_equal, self, other)
+        return self.tenfun(less_or_equal, other)
 
     def __gt__(self, other):
         """
@@ -2176,7 +2339,7 @@ class tensor:
         def greater(x, y):
             return x > y
 
-        return tt_tenfun(greater, self, other)
+        return self.tenfun(greater, other)
 
     def __lt__(self, other):
         """
@@ -2208,7 +2371,7 @@ class tensor:
         def less(x, y):
             return x < y
 
-        return tt_tenfun(less, self, other)
+        return self.tenfun(less, other)
 
     def __sub__(self, other):
         """
@@ -2240,7 +2403,7 @@ class tensor:
         def minus(x, y):
             return x - y
 
-        return tt_tenfun(minus, self, other)
+        return self.tenfun(minus, other)
 
     def __add__(self, other):
         """
@@ -2275,7 +2438,7 @@ class tensor:
         def tensor_add(x, y):
             return x + y
 
-        return tt_tenfun(tensor_add, self, other)
+        return self.tenfun(tensor_add, other)
 
     def __radd__(self, other):
         """
@@ -2325,7 +2488,7 @@ class tensor:
         def tensor_pow(x, y):
             return x**y
 
-        return tt_tenfun(tensor_pow, self, power)
+        return self.tenfun(tensor_pow, power)
 
     def __mul__(self, other):
         """
@@ -2360,7 +2523,7 @@ class tensor:
         if isinstance(other, (ttb.ktensor, ttb.sptensor, ttb.ttensor)):
             other = other.full()
 
-        return tt_tenfun(mul, self, other)
+        return self.tenfun(mul, other)
 
     def __rmul__(self, other):
         """
@@ -2418,7 +2581,7 @@ class tensor:
             with np.errstate(divide="ignore", invalid="ignore"):
                 return x / y
 
-        return tt_tenfun(div, self, other)
+        return self.tenfun(div, other)
 
     def __rtruediv__(self, other):
         """
@@ -2449,7 +2612,7 @@ class tensor:
             with np.errstate(divide="ignore", invalid="ignore"):
                 return x / y
 
-        return tt_tenfun(div, other, self)
+        return self.tenfun_binary(div, other, first=False)
 
     def __pos__(self):
         """
