+++ tests/test_pyttb_utils.py
@@ -75,59 +75,6 @@ def test_tt_dimscheck():
     assert "Negative dims" in str(excinfo), f"{str(excinfo)}"
 
 
-def test_tt_tenfun():
-    data = np.array([[1, 2, 3], [4, 5, 6]])
-    t1 = ttb.tensor(data)
-    t2 = ttb.tensor(data)
-
-    # Binary case
-    def add(x, y):
-        return x + y
-
-    assert np.array_equal(ttb_utils.tt_tenfun(add, t1, t2).data, 2 * data)
-
-    # Single argument case
-    def add1(x):
-        return x + 1
-
-    assert np.array_equal(ttb_utils.tt_tenfun(add1, t1).data, (data + 1))
-
-    # Multi argument case
-    def tensor_max(x):
-        return np.max(x, axis=0)
-
-    assert np.array_equal(ttb_utils.tt_tenfun(tensor_max, t1, t1, t1).data, data)
-    # TODO: sptensor arguments, depends on fixing the indexing ordering
-
-    # No np array case
-    assert np.array_equal(ttb_utils.tt_tenfun(tensor_max, data, data, data).data, data)
-
-    # No argument case
-    with pytest.raises(AssertionError) as excinfo:
-        ttb_utils.tt_tenfun(tensor_max)
-    assert "Must provide element(s) to perform operation on" in str(excinfo)
-
-    # No list case
-    with pytest.raises(AssertionError) as excinfo:
-        ttb_utils.tt_tenfun(tensor_max, [1, 2, 3])
-    assert "Invalid input to ten fun" in str(excinfo)
-
-    # Scalar argument not in first two positions
-    with pytest.raises(AssertionError) as excinfo:
-        ttb_utils.tt_tenfun(tensor_max, t1, t1, 1)
-    assert "Invalid input to ten fun" in str(excinfo)
-
-    # Tensors of different sizes
-    with pytest.raises(AssertionError) as excinfo:
-        ttb_utils.tt_tenfun(
-            tensor_max,
-            t1,
-            t1,
-            ttb.tensor(np.concatenate((data, np.array([[7, 8, 9]])))),
-        )
-    assert "Tensor 2 is not the same size as the first tensor input" in str(excinfo)
-
-
 def test_tt_setdiff_rows():
     a = np.array([[4, 6], [1, 9], [2, 6], [2, 6], [99, 0]])
     b = np.array(
